change length or size specifiers in APIs from int to long add array API (http://cr.yp.to/lib/array.html)

+++ byte/byte_chr.c
@@ -2,7 +2,7 @@
 
 /* byte_chr returns the smallest integer i between 0 and len-1
  * inclusive such that one[i] equals needle, or len it not found. */
-unsigned int byte_chr(const void* haystack, unsigned int len, char needle) {
+unsigned long byte_chr(const void* haystack, unsigned long len, char needle) {
   register char c=needle;
   register const char* s=haystack;
   register const char* t=s+len;
