+++ fmt/fmt_fill.c
@@ -4,8 +4,8 @@
  * append padlen-srclen spaces after dest, if that is >= 0.  Truncate
  * only if total length is larger than maxlen.  Return number of
  * characters written. */
-unsigned int fmt_fill(char* dest,unsigned int srclen,unsigned int padlen,unsigned int maxlen) {
-  int todo;
+unsigned long fmt_fill(char* dest,unsigned long srclen,unsigned long padlen,unsigned long maxlen) {
+  long todo;
   char* olddest=dest;
   char* max=dest+maxlen;
   if (dest==0) {
