+++ fmt/fmt_strn.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_strn(char *out,const char *in,unsigned int limit) {
+unsigned long fmt_strn(char *out,const char *in,unsigned long limit) {
   register char* s=out;
   register const char* t=in;
   register const char* u=in+limit;
