+++ scan/scan_nonwhitenskip.c
@@ -1,7 +1,7 @@
 #include "scan.h"
 #include <ctype.h>
 
-unsigned int scan_nonwhitenskip(const char *s,unsigned int limit) {
+unsigned long scan_nonwhitenskip(const char *s,unsigned long limit) {
   register const char *t=s;
   register const char *u=t+limit;
   while (t<u && !isspace(*t)) ++t;
